Restore readme caching suggestion
While this action is selective about installing as few dependencies as
are necessary for running ESLint, ESLint itself has quite a few
dependencies, and if you add more plugins to the mix, even the minimal
install can be sizable. In some (many?) cases it could still be
beneficial to cache what's installed. This is what used to live in the
suggested config, with a note about its optionality.